# contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: Archon API
  version: "1.0.0"
  description: Enterprise AI Orchestration & Governance Platform
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: Primary API
security:
  - BearerAuth: []
paths:
...
  /agents:
    get:
      operationId: listAgents
      summary: List all agents
      tags:
        - agents
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      operationId: createAgent
      summary: Create a new agent
      tags:
        - agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCreate"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ────────────────────────────────────────────
  # Agents — single resource
  # ────────────────────────────────────────────
  /agents/{agent_id}:
    parameters:
      - $ref: "#/components/parameters/AgentIdParam"
    get:
      operationId: getAgent
      summary: Get a single agent by ID
      tags:
        - agents
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      operationId: updateAgent
      summary: Update an existing agent
      tags:
        - agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      operationId: deleteAgent
      summary: Delete an agent
      tags:
        - agents
      responses:
        "204":
          description: Agent deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ────────────────────────────────────────────
  # Executions — trigger
  # ────────────────────────────────────────────
  /execute:
    post:
      operationId: executeAgent
      summary: Execute an agent
      tags:
        - executions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecutionRequest"
      responses:
        "202":
          description: Execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ────────────────────────────────────────────
  # Executions — list & detail
  # ────────────────────────────────────────────
  /executions:
    get:
      operationId: listExecutions
      summary: List executions with pagination
      tags:
        - executions
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - name: agent_id
          in: query
          description: Filter executions by agent ID
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Filter executions by status
          schema:
            type: string
            enum: [queued, running, completed, failed, cancelled]
      responses:
        "200":
          description: List of executions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /executions/{execution_id}:
    parameters:
      - $ref: "#/components/parameters/ExecutionIdParam"
    get:
      operationId: getExecution
      summary: Get a single execution by ID
      tags:
        - executions
      responses:
        "200":
          description: Execution details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ────────────────────────────────────────────
  # Users
  # ────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by Keycloak. Obtain a token via the Keycloak OpenID Connect
        token endpoint (grant_type=authorization_code or client_credentials).
        Include as `Authorization: Bearer <token>`.
    KeycloakOAuth2:
      type: oauth2
      description: Keycloak OAuth 2.0 authorization code and client credentials flows.
      flows:
        authorizationCode:
          authorizationUrl: /realms/archon/protocol/openid-connect/auth
          tokenUrl: /realms/archon/protocol/openid-connect/token
          refreshUrl: /realms/archon/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: User profile information
            email: User email address
            roles: User role assignments
        clientCredentials:
          tokenUrl: /realms/archon/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            roles: User role assignments
    OpenIDConnect:
      type: openIdConnect
      description: Keycloak OpenID Connect discovery endpoint.
      openIdConnectUrl: /realms/archon/.well-known/openid-configuration

  # ────────────────────────────────────────────
  # Parameters
  # ────────────────────────────────────────────
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
    AgentIdParam:
      name: agent_id
      in: path
      required: true
      description: Unique agent identifier
      schema:
        type: string
        format: uuid
    ExecutionIdParam:
      name: execution_id
      in: path
      required: true
      description: Unique execution identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request — invalid or malformed input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "BAD_REQUEST"
                message: "Request body is malformed"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"
    Unauthorized:
      description: Authentication required — missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "UNAUTHORIZED"
                message: "Bearer token is missing or expired"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"
    Forbidden:
      description: Forbidden — insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "FORBIDDEN"
                message: "You do not have permission to access this resource"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "NOT_FOUND"
                message: "The requested resource was not found"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"
    UnprocessableEntity:
      description: Validation error — input is well-formed but semantically invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "VALIDATION_ERROR"
                message: "Field value is invalid"
                field: "name"
                details: "Name must be between 1 and 255 characters"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            errors:
              - code: "INTERNAL_ERROR"
                message: "An unexpected error occurred"
            meta:
              request_id: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2025-01-15T12:00:00Z"

  # ────────────────────────────────────────────
  # Schemas
  # ────────────────────────────────────────────
  schemas:
...
    ResponseMeta:
      type: object
      required:
        - request_id
        - timestamp
      properties:
        request_id:
          type: string
          format: uuid
          description: Unique identifier for this request (correlation ID)
        timestamp:
          type: string
          format: date-time
          description: Server timestamp when the response was generated
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
    PaginationMeta:
      type: object
      required:
        - total
        - limit
        - offset
        - has_more
      properties:
        total:
          type: integer
          description: Total number of items matching the query
          minimum: 0
        limit:
          type: integer
          description: Maximum items per page
          minimum: 1
          maximum: 100
        offset:
          type: integer
          description: Number of items skipped
          minimum: 0
        has_more:
          type: boolean
          description: Whether more items exist beyond this page
    ErrorEnvelope:
      type: object
      required:
        - errors
        - meta
      properties:
        errors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ErrorDetail"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples:
            - "BAD_REQUEST"
            - "VALIDATION_ERROR"
            - "NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
        field:
          type: string
          description: Name of the field that caused the error (for validation errors)
        details:
          type: string
          description: Additional context about the error

    # ── Health ────────────────────────────────
...
    AgentCreate:
      type: object
      required:
        - name
        - definition
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Human-readable agent name
        description:
          type: string
          maxLength: 2000
          description: Optional description of the agent's purpose
        definition:
          type: object
          description: JSON graph definition describing the agent's node topology and edges
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
            maxLength: 64
          maxItems: 20
          description: Optional tags for categorisation and search
        template_id:
          type: string
          format: uuid
          description: Optional ID of the template this agent was created from
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Human-readable agent name
        description:
          type: string
          maxLength: 2000
          description: Optional description of the agent's purpose
        definition:
          type: object
          description: JSON graph definition describing the agent's node topology and edges
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
            maxLength: 64
          maxItems: 20
          description: Optional tags for categorisation and search
    Agent:
      type: object
      required:
        - id
        - name
        - definition
        - status
        - owner_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique agent identifier
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Human-readable agent name
        description:
          type: string
          maxLength: 2000
          description: Description of the agent's purpose
        definition:
          type: object
          description: JSON graph definition (nodes, edges, parameters)
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
          description: Tags for categorisation and search
        template_id:
          type: string
          format: uuid
          description: ID of the template this agent was created from (if any)
        status:
          type: string
          enum: [draft, active, archived]
          description: Current lifecycle status of the agent
        owner_id:
          type: string
          format: uuid
          description: ID of the user who owns this agent
        created_at:
          type: string
          format: date-time
          description: Timestamp when the agent was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp of the last update
    AgentResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    AgentListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/ResponseMeta"

    # ── Execution ─────────────────────────────
    ExecutionRequest:
      type: object
      required:
        - agent_id
      properties:
        agent_id:
          type: string
          format: uuid
          description: ID of the agent to execute
        input:
          type: object
          additionalProperties: true
          description: Input data to pass to the agent
        parameters:
          type: object
          additionalProperties: true
          description: Runtime parameters (timeouts, retries, model overrides, etc.)
        callback_url:
          type: string
          format: uri
          description: Optional URL to POST results to when execution completes
    Execution:
      type: object
      required:
        - id
        - agent_id
        - status
        - owner_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique execution identifier
        agent_id:
          type: string
          format: uuid
          description: ID of the agent that was executed
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
          description: Current execution status
        input:
          type: object
          additionalProperties: true
          description: Input data that was provided
        output:
          type: object
          additionalProperties: true
          description: Output data produced by the execution (null while running)
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
          description: Error information (only present when status is failed)
        parameters:
          type: object
          additionalProperties: true
          description: Runtime parameters used for this execution
        owner_id:
          type: string
          format: uuid
          description: ID of the user who triggered the execution
        started_at:
          type: string
          format: date-time
          description: Timestamp when execution began running (null if still queued)
        completed_at:
          type: string
          format: date-time
          description: Timestamp when execution reached a terminal state (null if still running)
        duration_ms:
          type: integer
          minimum: 0
          description: Execution duration in milliseconds (null if not completed)
        created_at:
          type: string
          format: date-time
          description: Timestamp when the execution was created (queued)
        updated_at:
          type: string
          format: date-time
          description: Timestamp of the last status change
    ExecutionResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: "#/components/schemas/Execution"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ExecutionListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Execution"
        meta:
          $ref: "#/components/schemas/ResponseMeta"

    # ── User ──────────────────────────────────
...